Get GeoCode Before Create City

+++ IoTManager.Model/CityModel.cs
@@ -11,8 +11,8 @@ namespace IoTManager.Model
         public String Remark { get; set; }
         public DateTime CreateTime { get; set; }
         public DateTime UpdateTime { get; set; }
-        public int longitude{ get; set; }
-        public int latitude{ get; set; }
+        public double longitude{ get; set; }
+        public double latitude{ get; set; }
 
     }
 }
